ZOOKEEPER-3193. Refactor SaslAuthFail test to use single class. Use CountDownLatch to sync with watcher.
